requirements.txt
flask==2.0.3
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
numpy
python-dotenv
opencv-python
Pillow